Refactor Sources and Cloud provider APIs
