+++ api/openapi.gen.yaml
@@ -210,6 +210,7 @@ paths:
                 $ref: '#/components/schemas/v1.AWSReservationResponse'
         "500":
           $ref: '#/components/responses/InternalError'
+  /reservations/aws/{ID}:
     get:
       description: 'Return an AWS reservation with details by id'
       operationId: getAWSReservationByID
